requirements.txt
aiohttp~=3.9.5
requests~=2.31.0
python-dotenv~=1.0.1
deep-translator==1.11.4
python-telegram-bot==21.0.1